Colocate render methods with their types
